// back-end/src/types/doacaoAlimentoTypes.ts
export type DoacaoAlimentoCreateType = {
    id_solicitante: number;
    id_alimento: number;
    peso: number;
    id_condutor: number;
    id_estoque_alimento: number;
};

export type DoacaoAlimentoUpdateType = {
    id_solicitante: number;
    id_alimento: number;
    peso: number;
    id_condutor: number;
    id_estoque_alimento: number;
};